serverless_functions/image-upload/requirements.txt
minio==7.1.15
kafka-python==2.0.2
pillow==9.5.0 